Create gh-8073-bitcraft.json
I added some lore to the location of this artpiece. I was involved in these discussions and the relocation of the logo from the VTubers Place's side and I wanted to add some more information about the placement of the logo.
Basically I just write down what I think would be interesting to read in a year, that you otherwise would've completely forgotten.